+++ lib/capistrano/slacky.rb
@@ -2,6 +2,8 @@
 
 load File.expand_path("../../tasks/slacky.rake", __FILE__)
 
+require "forwardable"
+
 require_relative "slacky/version"
 require_relative "slacky/configuration"
 require_relative "slacky/i18n"
@@ -14,17 +16,9 @@ module Capistrano
   module Slacky
     module_function
 
-    # Delegate any missing method call to the configuration
-    def method_missing(method_name, *arguments, &block)
-      return super unless configuration.respond_to?(method_name)
+    extend ::SingleForwardable
 
-      configuration.public_send(method_name, *arguments, &block)
-    end
-
-    # Replace the Object.respond_to?() method
-    def respond_to_missing?(method_name, include_private = false)
-      configuration.respond_to?(method_name) || super
-    end
+    def_delegators :configuration, :username, :icon_emoji, :channel, :klass, :slacky?, :repo
 
     # @return [Capistrano::Slacky::Configuration]
     def configuration