+++ CHANGELOG.md
@@ -5,9 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-###
-Security
+### Security
 - Fix possible command injection during the duration retrieval. ([@chubchenko][])
+
+### Changed
+- Use `I18n.t` instead of `t`. ([@chubchenko][])
 
 ## [0.1.1] - 2021-05-27
 ### Fixed