+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Use `I18n.t` instead of `t`. ([@chubchenko][])
 - Use `forwardable` instead of `method_missing` + `respond_to_missing?`. ([@chubchenko][])
+- Use the `On` module to interact with the remote server. ([@chubchenko][])
 
 ### Fixed
 - Use `@env` instance variable instead of missing getter `env`. ([@chubchenko][])