+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.1.2] - 2021-06-03
 ### Changed
 - Use `I18n.t` instead of `t`. ([@chubchenko][])
 - Use `forwardable` instead of `method_missing` + `respond_to_missing?`. ([@chubchenko][])
@@ -32,5 +32,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [@chubchenko]: https://github.com/chubchenko
 [Unreleased]: https://github.com/chubchenko/capistrano-slacky/compare/v0.1.1...HEAD
+[0.1.2]: https://github.com/chubchenko/capistrano-slacky/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/chubchenko/capistrano-slacky/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/chubchenko/capistrano-slacky/releases/tag/v0.1.0